v2

+++ URLShortener.Business/Services/LinkService.cs
@@ -1,15 +1,11 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
+﻿using System.Text;
 using URLShortener.Business.Interfaces;
 using URLShortener.Data.Entities;
 using URLShortener.Data.Interfaces;
 
 namespace URLShortener.Business.Services
 {
-    public class LinkService: ILinkService
+    public class LinkService : ILinkService
     {
         private readonly ILinkRepository _linkRepository;
         private readonly ITokenService _tokenService;
@@ -22,37 +18,75 @@ namespace URLShortener.Business.Services
 
         public async Task CreateLinkAsync(string originalLink)
         {
+            if (await _linkRepository.GetLinkByOriginalURLAsync(originalLink) != null)
+            {
+                throw new InvalidDataException("Link already exists");
+            }
             var link = new Link
             {
                 OriginalURL = originalLink,
                 ShortenedURL = GenerateShortCode(originalLink),
-                UserId = _tokenService.GetUserIdFromToken()
+                UserId = _tokenService.GetUserIdFromToken(),
+                CreatedAt = DateTime.Now
             };
             await _linkRepository.CreateLinkAsync(link);
         }
 
-        private string GenerateShortCode(string originalUrl)
-        {
-            byte[] hashBytes = System.Security.Cryptography.MD5.Create().ComputeHash(Encoding.UTF8.GetBytes(originalUrl));
-            string base64String = Convert.ToBase64String(hashBytes);
-            base64String.TrimEnd('=').Replace("/", "_").Replace("+", "-");
-
-            return base64String;
-        }
-
         public async Task<string> GetOriginalLinkByShortenedURLAsync(string shortenedURL)
         {
-            return await _linkRepository.GetOriginalLinkByShortenedURLAsync(shortenedURL);
+            var originalLink = await _linkRepository.GetOriginalLinkByShortenedURLAsync(shortenedURL);
+
+            if (originalLink == null)
+            {
+                throw new NullReferenceException("Link not found");
+            }
+
+            return originalLink;
         }
 
         public async Task<Link> GetLinkByIdAsync(int id)
         {
-            return await _linkRepository.GetLinkByIdAsync(id);
+            var link = await _linkRepository.GetLinkByIdAsync(id);
+
+            if (link == null)
+            {
+                throw new NullReferenceException("Link not found");
+            }
+
+            return link;
+        }
+
+        public async Task<IEnumerable<Link>> GetAllLinksAsync()
+        {
+            return await _linkRepository.GetAllLinksAsync();
         }
 
         public async Task DeleteLinkByIdAsync(int id)
         {
-            await _linkRepository.DeleteLinkByIdAsync(id);
+            var link = await _linkRepository.GetLinkByIdAsync(id);
+
+            if (link == null)
+            {
+                throw new NullReferenceException("Link not found");
+            }
+
+            if (_tokenService.GetUserIdFromToken() == link.UserId || _tokenService.IsAdminFromToken())
+            {
+                await _linkRepository.DeleteLinkByIdAsync(id);
+            }
+            else
+            {
+                throw new UnauthorizedAccessException("You don't have permisssion to delete this link");
+            }
+        }
+
+        private string GenerateShortCode(string originalUrl)
+        {
+            byte[] hashBytes = System.Security.Cryptography.SHA256.Create().ComputeHash(Encoding.UTF8.GetBytes(originalUrl));
+            string base64String = Convert.ToBase64String(hashBytes);
+            base64String = base64String.TrimEnd('=').Replace("/", "_").Replace("+", "-");
+
+            return base64String.Substring(0, 8);
         }
     }
 }
